add: マイグレーションの練習用コード
