--- ops/1_/migrate-demo/prisma/schema.prisma
@@ -1,24 +1,0 @@
-// This is your Prisma schema file,
-// learn more about it in the docs: https://pris.ly/d/prisma-schema
-
-generator client {
-  provider = "prisma-client-js"
-}
-
-datasource db {
-  provider = "postgresql"
-  url      = env("DATABASE_URL")
-}
-
-model Users {
-  id     String @id @default(uuid())
-  name   String
-  Pairs  Pairs  @relation(fields: [pairId], references: [id])
-  pairId String
-}
-
-model Pairs {
-  id    String  @id @default(uuid())
-  name  String  @db.Char(1)
-  Users Users[]
-}
