+++ ops/2_本番稼働中のデータベースをマイグレーションしよう/migrate-demo/tsconfig.json
